pyproject.toml
[tool.poetry]
name = "mfs"
version = "0.1.0"
description = ""
authors = ["Grigoris Bourdalas <bourdalasgr@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.6"
fastapi = "^0.95"
uvicorn = "^0.21"
streamlit = "^1.21"
streamlit_pydantic = "^0.6"
streamlit-extras = "^0.2.7" 
streamlit-authenticator = "^0.2.1"
streamlit-autorefresh = "^0.0.1"
pyjwt = "^2.6.0"
sqlalchemy = "^2.0.9"
dictdiffer = "^0.9.0" 
pyngrok = "^5.2.1"
python-jose = "^3.0.2"
geopandas = "^0.12.2"
plotly = "^5.14.1"
pydeck = "^0.8.0"
geocoder = "^1.38.1"
geopy = "^2.3.0"
pycountry = "^22.3.5"
python-dotenv = "^1.0.0"


[tool.poetry.extras]
email = ["pydantic"]

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
